requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.108.0
fastapi_camelcase==1.0.5
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.14.6
pyhumps==3.8.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqladmin==0.15.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
starlette==0.32.0.post1
typing_extensions==4.9.0
uuid6==2023.5.2
uvicorn==0.25.0
WTForms==3.1.1
